+++ application/src/main/java/uk/gov/hmcts/dm/config/AuditorAwareImpl.java
@@ -1,0 +1,23 @@
+package uk.gov.hmcts.dm.config;
+
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import uk.gov.hmcts.reform.auth.checker.spring.serviceonly.ServiceDetails;
+
+import java.util.Optional;
+
+public class AuditorAwareImpl implements AuditorAware<String> {
+
+    @Override
+    public Optional<String> getCurrentAuditor() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            ServiceDetails userDetails = (ServiceDetails) authentication.getPrincipal();
+            return Optional.ofNullable(userDetails.getUsername());
+        } else {
+            return Optional.empty();
+        }
+    }
+
+}
