+++ application/src/main/java/uk/gov/hmcts/dm/controller/DocumentContentVersionController.java
@@ -31,7 +31,6 @@ import uk.gov.hmcts.dm.service.StoredDocumentService;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.io.IOException;
-import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -125,36 +124,36 @@ public class DocumentContentVersionController {
         @PathVariable UUID versionId,
         HttpServletResponse response) {
 
-            DocumentContentVersion documentContentVersion = documentContentVersionService.findById(versionId)
-                .orElseThrow(() -> new DocumentContentVersionNotFoundException(versionId));
+        DocumentContentVersion documentContentVersion = documentContentVersionService.findById(versionId)
+            .orElseThrow(() -> new DocumentContentVersionNotFoundException(versionId));
 
-            if (documentContentVersion.getStoredDocument().isDeleted()) {
-                throw new DocumentContentVersionNotFoundException(versionId);
+        if (documentContentVersion.getStoredDocument().isDeleted()) {
+            throw new DocumentContentVersionNotFoundException(versionId);
+        }
+
+        response.setHeader(HttpHeaders.CONTENT_TYPE, documentContentVersion.getMimeType());
+        response.setHeader(HttpHeaders.CONTENT_LENGTH, documentContentVersion.getSize().toString());
+        response.setHeader("OriginalFileName", documentContentVersion.getOriginalDocumentName());
+        response.setHeader(HttpHeaders.CONTENT_DISPOSITION,
+            String.format("fileName=\"%s\"", documentContentVersion.getOriginalDocumentName()));
+
+        try {
+            if (isBlank(documentContentVersion.getContentUri())) {
+                response.setHeader("data-source", "Postgres");
+                auditedDocumentContentVersionOperationsService.readDocumentContentVersionBinary(documentContentVersion, response.getOutputStream());
+            } else {
+                response.setHeader("data-source", "contentURI");
+                auditedDocumentContentVersionOperationsService.readDocumentContentVersionBinaryFromBlobStore(
+                    documentContentVersion,
+                    response.getOutputStream());
             }
+            response.flushBuffer();
 
-            response.setHeader(HttpHeaders.CONTENT_TYPE, documentContentVersion.getMimeType());
-            response.setHeader(HttpHeaders.CONTENT_LENGTH, documentContentVersion.getSize().toString());
-            response.setHeader("OriginalFileName", documentContentVersion.getOriginalDocumentName());
-            response.setHeader(HttpHeaders.CONTENT_DISPOSITION,
-                String.format("fileName=\"%s\"", documentContentVersion.getOriginalDocumentName()));
-
-            try {
-                if (isBlank(documentContentVersion.getContentUri())) {
-                    response.setHeader("data-source", "Postgres");
-                    auditedDocumentContentVersionOperationsService.readDocumentContentVersionBinary(documentContentVersion, response.getOutputStream());
-                } else {
-                    response.setHeader("data-source", "contentURI");
-                    auditedDocumentContentVersionOperationsService.readDocumentContentVersionBinaryFromBlobStore(
-                        documentContentVersion,
-                        response.getOutputStream());
-                }
-                response.flushBuffer();
-
-            } catch (IOException e) {
-                return ResponseEntity
-                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(e);
-            }
+        } catch (IOException e) {
+            return ResponseEntity
+                .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                .body(e);
+        }
 
         return null;
     }
