+++ application/src/main/java/uk/gov/hmcts/dm/service/BlobStorageMigrationService.java
@@ -29,7 +29,6 @@ import java.sql.Blob;
 import java.sql.SQLException;
 import java.time.Duration;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 import static java.lang.System.currentTimeMillis;
@@ -68,11 +67,6 @@ public class BlobStorageMigrationService {
         this.auditEntryRepository = auditEntryRepository;
         this.batchMigrationTokenService = batchMigrationTokenService;
         this.batchMigrationAuditEntryService = batchMigrationAuditEntryService;
-    }
-
-    public void migrateDocumentContentVersion(@NotNull UUID documentId, @NotNull UUID versionId) {
-        final DocumentContentVersion documentContentVersion = getDocumentContentVersion(documentId, versionId);
-        migrateDocumentContentVersion(documentContentVersion);
     }
 
     public BatchMigrateProgressReport batchMigrate(String authToken,
@@ -115,6 +109,11 @@ public class BlobStorageMigrationService {
     public MigrateProgressReport getMigrateProgressReport() {
         return new MigrateProgressReport(documentContentVersionRepository.countByContentChecksumIsNull(),
                                          documentContentVersionRepository.countByContentChecksumIsNotNull());
+    }
+
+    public void migrateDocumentContentVersion(@NotNull UUID documentId, @NotNull UUID versionId) {
+        final DocumentContentVersion documentContentVersion = getDocumentContentVersion(documentId, versionId);
+        migrateDocumentContentVersion(documentContentVersion);
     }
 
     private void migrateDocumentContentVersion(DocumentContentVersion documentContentVersion) {
