+++ application/src/test/java/uk/gov/hmcts/dm/controller/DocumentThumbnailControllerTests.java
@@ -13,7 +13,6 @@ import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 import java.util.UUID;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
